powernode/dim/pattern_engines/collaborative/requirements.txt
# Collaborative Pattern Engine Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
pyyaml>=6.0.1

# IPFS integration (uses powernode.ipfs)
requests>=2.28.0

